Move LookupImpl to another file.
